refactor: refactor caption for consistency
