+++ packages/caption/src/caption-prop-types.js
@@ -1,0 +1,22 @@
+import { Text, ViewPropTypes } from "react-native";
+import PropTypes from "prop-types";
+
+const { style: TextPropTypesStyle } = Text.propTypes;
+const { style: ViewPropTypesStyle } = ViewPropTypes;
+
+export const propTypes = {
+  children: PropTypes.element,
+  credits: PropTypes.string,
+  style: PropTypes.shape({
+    text: TextPropTypesStyle,
+    container: ViewPropTypesStyle
+  }),
+  text: PropTypes.string
+};
+
+export const defaultProps = {
+  children: null,
+  credits: null,
+  style: {},
+  text: null
+};
